backend/requirements.txt
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
python-multipart==0.0.6
pymongo==4.5.0
motor==3.3.1

